+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - LR Parser
     - Add `ParserBuilder.use` to re-use existing parser builder or modularize code.
     - `GrammarRule/Candidate.toString` will use `<=` as the default arrow instead of `=>`.
+- Optimize document / comments.
 
 ## v0.2.0